Fixed potential bugs while output csv and markdown; Moved some low couple functions to util.ts; Refactored core functions for convenience;
